Renamed _socket back to socket to try if it was a GH problem
