+++ sentry_sdk/integrations/socket.py
